requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
anthropic>=0.39.0
openai>=2.0.0
pandas==2.2.0
python-dotenv==1.0.1
